Rename config.py -> configuration.py
This is to not cause conflicts for those that have already set up their
environments.
