+++ disc.py
@@ -2,7 +2,7 @@ from discord.ext.commands import Command, Context
 
 from logger import logger
 
-from config import config
+from configuration import config
 
 __all__ = ["DiscordCommand"]
